Create separate class for combining two dense addresses.
